add --all option to scripts
