+++ typescript/infra/scripts/agent-utils.ts
@@ -26,6 +26,7 @@ import { Contexts } from '../config/contexts.js';
 import { agents } from '../config/environments/agents.js';
 import { WarpRouteIds } from '../config/environments/mainnet3/warp/warpIds.js';
 import { validatorBaseConfigsFn } from '../config/environments/utils.js';
+import { AlertType } from '../config/grafanaAlerts.js';
 import {
   getChain,
   getChainAddresses,
@@ -55,7 +56,6 @@ import {
 } from '../src/utils/utils.js';
 
 import { BalanceThresholdType } from './funding/utils/constants.js';
-import { AlertType } from './funding/utils/grafana.js';
 
 const debugLog = rootLogger.child({ module: 'infra:scripts:utils' }).debug;
 
@@ -224,6 +224,13 @@ export function withAlertType<T>(args: Argv<T>) {
 
 export function withAlertTypeRequired<T>(args: Argv<T>) {
   return withAlertType(args).demandOption('alertType');
+}
+
+export function withConfirmAllChoices<T>(args: Argv<T>) {
+  return args
+    .describe('all', 'Confirm all choices')
+    .boolean('all')
+    .default('all', false);
 }
 
 export function withAgentRole<T>(args: Argv<T>) {
