+++ typescript/infra/scripts/funding/read-alert-thresholds.ts
@@ -2,14 +2,14 @@ import yargs from 'yargs';
 
 import { rootLogger } from '@hyperlane-xyz/utils';
 
+import { alertConfigMapping } from '../../config/grafanaAlerts.js';
 import { writeJsonAtPath } from '../../src/utils/utils.js';
 import { withAlertTypeRequired, withWrite } from '../agent-utils.js';
 
 import {
   THRESHOLD_CONFIG_PATH,
-  alertConfigMapping,
   getAlertThresholds,
-  orderThresholds,
+  sortThresholds,
 } from './utils/grafana.js';
 
 async function main() {
@@ -18,22 +18,16 @@ async function main() {
   ).argv;
 
   const alertThresholds = await getAlertThresholds(alertType);
-  const orderedThresholds = orderThresholds(alertThresholds);
+  const sortedThresholds = sortThresholds(alertThresholds);
 
-  const alertThresholdArray = Object.entries(orderedThresholds).map(
-    ([chain, threshold]) => ({
-      chain,
-      threshold,
-    }),
-  );
-  console.table(alertThresholdArray);
+  console.table(sortedThresholds);
 
   if (write) {
     rootLogger.info('Writing alert thresholds to file..');
     try {
       writeJsonAtPath(
         `${THRESHOLD_CONFIG_PATH}/${alertConfigMapping[alertType].configFileName}`,
-        orderedThresholds,
+        sortedThresholds,
       );
       rootLogger.info('Alert thresholds written to file.');
     } catch (e) {
